% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{committee_code}
\alias{committee_code}
\title{Committee Type Codes}
\format{A data frame with two variables:
\describe{
\item{\code{committee_code}}{A 1-character alphabetic code assigned to a committee by the Federal Election Commission. The committee type code for a specific committee always remains the same.}
\item{\code{committee_type}}{Type of committee as recognized by the FEC.}
}

For further details and source, see \url{http://www.fec.gov/finance/disclosure/metadata/CommitteeTypeCodes.shtml}}
\usage{
committee_code
}
\description{
Committee Type Codes
}
\keyword{datasets}

