% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{party_code}
\alias{party_code}
\title{Party Codes}
\format{A data frame with two variables:
\describe{
\item{\code{party_code}}{A 1-3 character alphabetic code assigned to a party by the Federal Election Commission. The party code for a specific party always remains the same.}
\item{\code{party}}{Official name of party as recognized by the FEC.}
}

For further details and source, see \url{http://www.fec.gov/finance/disclosure/metadata/DataDictionaryPartyCodeDescriptions.shtml}}
\usage{
party_code
}
\description{
Party Codes
}
\keyword{datasets}

